Implemented a resolver cache so that hostname lookups occur at most once.
The purpose of the cache is primarily to defeat DNS round-robin
schemes. It has the added advantage of minimizing the number of DNS
query packets transmitted over the network. Use of the cache can be
disabled on a per-request basis.

The resolver cache is currently used only for End-to-End transfers.