Remove the MSG_WAITALL flag from recvfrom(2).
Deadlock can occur when the MSG_WAITALL flag is specified when reading
data from a socket. This patch removes the MSG_WAITALL flag and
restructures the code that depended on the size of the received
message.

The patch essentially converts the application-level protocol for
end-to-end transfers to use fixed size messages. All messages
(including the end-of-file message) are padded to the request size
(i.e. p->iosize).