Changes to add feature based configure options in config.h, modularization for the target offset table, and the inclusion of the include guards for some of the global data headers.
The config.h change moves from haveing HAVE_* features defined in the makefile
to having those in a config file.  This should make better cross-platform
support easier.  Specifically, the target offset table tries to use valloc
or posix_memalign if those are available.

The target offset table has been made more modular with that addition of
init, destroy, and update functions in target_offset_table.c.  This is the
beginning of a long push to get bookkeeping structure updates and program flow
separated to the appropriate degree.  This will not ever be perfect, and nor
does it need to be, but this at least allows for easier initialization and
destruction of the TOT.

Finally, in trying to leverage as few includes as possible, I found a few
unprotected headers, and added include guards for those.  Minor changes moving
towards a reduced dependency DAG in the future.