Fix all CLang Warnings
This kills all the warnings emitted by Clang.  Mainly,
the remaining warnings are caused by mismatched signs.  Some
of these lead to true undefined behavior because:
  a)  Stat returns the file size as an off_t
  b)  Read takes an io_size of type size_t
  c)  Read returns the successful io_size as a ssize_t

So, in the long run, we need to check and make sure that we
never read or write buffers larger than SSIZE_MAX.  This
is pretty unlikely, but probably leads to problems if it
happens.