Fix small diverse build warnings
Found some duplicated files in the XNI dir and removed those.
Also fixed a compile issue with an in_addr_t not being initialized.
Initialized it to INADDR_LOOPBACK, and made werror not exactly the
default, but requiring a --enable-debug flag during configure.
Why not simply do CFLAGS=-werror ./configure?  Because the
configure process can't pass with that.  Eventually I suppose
we should set NDEBUG as part of that as well.