XNI-IB Compiles
This returns XNI to a state where the XNI-IB driver compiles.
Addmittedly this is because I commented huge chunks of it out,
though that was undoubtely better than before when it was skipped
by the preprocessor.  Also fixes the config file and Makefile
so that IB support can be autoconf'ed correctly.