Add distribution make targets
This commit removes ./configure from source control.  Hopefully,
this makes the commit messages easier to read.  As part of that
you will now have to run autoconf after cloning the code.  That
should be pretty easy as we allow really old versions of autoconf
to work with this source.  On the other hand, this makes the
release process trickier.  So I built a set of targets that match
the GNU requirements (more or less I think) for make dist and
make distcheck.  I in fact use git to perform the dist.  Thus
tar ball recipients will not be able to run distcheck themselves.
I think that is probably okay.  If the code moves to github I
may need to re-think this decision.  But if that happens, I
think we will host the releases at ioperformance.com, xdd.org,
or deepcopy.  So it should be fine.