Fix SUBTLE initstate issue
The initstate function is used to set the state of the random
number generator.  The function signature is:

char* initstate(int seed, char* state, size_t statesize)

The problems keys on what init state returns: the previous
state.  What this means is that during each call to initstate,
the memory allocated for the previous state must still exist.

In our case, each target calls initstate, however, each target
does it like this:

char state[256];
initstate(sp->seed, state, 256);
return;

And thus the value returned by initstate is a memory region
that was previously freed.  And this would be fine, except that
initstate *writes* to the old state before returning it to you!
This makes sense given a few hours of reflection, but it is a
hard to handle interface!  So I moved the states into the
seekhdr, which is stored on the heap, and I maintain the oldstate:

struct seekhdr {
  char state[256];
  char *oldstate;
}

sp->oldstate = initstate(sp->seed, sp->state, sizeof(sp->state));

In truth we should restore the old state with setstate, when the
seekhdr is freed, but we would actually restore it incorrectly,
since we would restore it based on target finish times.  So for
now I haven't added the setstate(sp->oldstate) call anywhere yet.

In short, when the return value is something weird like a pointer,
be extra extra careful.