XDD-Lite and libxdd work
This commit allows plans and targets to be created, started, and
waited on by libxdd.  Further, it fixes some issues with
xdd-lite that prevented the correct construction of xdd plans.

Libxdd is based on the principle that plans and targets should
never be in a partially constructed state.  So in order to
create a valid plan the user must first:

Initialize and create a set of target attributes
Initialize and create a plan attributes

These can be filled out in any order, as these are just attributes.
The user then uses these to construct a plan, which ensures
that all of the relevant data exists before libxdd begins
constructing the internal plan, target, and worker data structures.
This is extremely similar to how pthreads and xni connections are
safely created (basically any stateful interface).

Once the plan is created, the user can then start the plan (which
doesn't block) and then wait for the plan to complete.

Things that still don't work: e2e targets, and more esoteric options
not exposed by libxdd.