Modify preallocate for 32 vs 64 bit
In the past we only preallocated in 32bit (4GB) chunks due to
some occasional issues on 32 bit platforms.  This patch detects the
platform and allocates faster dependent on which platform.  The
patch still needs a little refinement as its a compile time
check that I'm performing at runtime.  I just haven't figured out
what the right way to detect the pointer size is yet.