Making building xdd more exacting
Before now building xdd was relatively easy; if you don't have a
development package, we work around that by default.  This is the
first in a likely series of minor fixes to make that harder.
The new default is --enable-ib --enable-numa and --enable-xfs.
If you don't want those features, or don't have the headers, or
are on a platform that cannot support those features, you will
have to disable the packages.  So just use
./configure --disable-ib --disable-numa --disable-xfs
and XDD will build.