Document why XFS extensions fail
XFS often fails to build not because of missing XFS packages, but
because both redhat and SuSE didn't realize that some older
versions of XFS relied on libuuid, and so the package dependency
is missing from the packages.  However, just checking for uuid
isn't a great idea because some versions of XFS don't require
uuid.  Here we just tell the user, that if they have XFS and the
code still doesn't configure, try adding uuid headers.