Make install also build XDD
Personally, I think this is a stupid change, and I don't like it.  But
the user base is struggling to build XDD and so they simply type
make install as root.  I don't support doing things that way.  In my
opinion, the correct way to build xdd is like this
./configure
make
sudo make install

or su -c if you are so inclined.  Why is it better?  Because once you
do this:
./configure
sudo make install

You can't run make clean.  You have to run sudo make clean.  Which
is, honestly dumb, and violates the idea behind LUP.  But whatever.
Automake works this way.  Dave Dillow likes it.  So whatever, I'm a
pragmatist, not an ideologue.  But I'm still writing this overly long
commit message to complain about it.