Create branch for XNI integration
This is the first checkin for a branch that can
be used to validate and work with a combination of
XNI and e2e.  This proof of concept branch is a staging area
to integrate this code into the mainline and to perform
V&V tasks with the XNI code.  The branch is expected to progress
as follows:

Step 1 - set up a switching mechanism to alternate between net and xnet
Step 2 - Add XNI init/finalize to code
Step 3 - Add XNI register/unregister to code
Step 4 - Add XNI take/release to code
Step 5 - COnvert to XNI send and receive code