Cleanup for public release
This updates some out of date documentation sprinkled in text
files and some older versions of the manual.  We now
accurately describe how to build XDD and describe some of the
current TODO items.  I also got rid of some of the old binary
versions of the manual.  In the long run we need a better scheme
for managing the manual, but this is a start.