XNI integration of xni_register_buffer (#18)
This checkin adds the creation of the xni control block and
context for each xni target, and it adds the calls to register
the buffer with the xni context.  This checkin also updates the
XNI documentation to correctly describe the parameters for the
xni_register_buffer call.  This should leave only the
establishment of the XNI connection and use of the XNI send/rx
primitives before we have working XNI support.