Add request/release calls for XNI (#18)
This commit adds the necessary request/release calls.  Since they
are done only in the networking code, it isn't clear that they will
be fast enough, or correctly allow all tabulation data.  In effect
I'm having to allow worker threads to change their I/O buffer
before each send/recv.  This shouldn't be a problem -- at least in
the single NIC case, but it is weird.  I suspect multi-NIC simply
doesn't work, so it's less of an issue there right now.  But will
need to eventually get that working.