Update xddmcp transfer manager.
This updates the function names to actually match what the functions
currently do in the TransferManager.  It also improves the monitoring
logic so that if an xdd process fails, the transfer is immediately
failed.  There is an argument not to do it this way, but if we
don't fail pessimistically it isn't possible to avoid an infinite
loop in some cases (as currently coded).