XDDMCP updates for multi-flows
This is the initial checkin that enables multi-host transfers.  It also fixes some
of the naming.  FlowBuilder doesn't describe the objects very well.  They are more
accurately, endpoints, but the cost of making the change globally is too high.  So
for now, fixing the higher levels, but leaving the guts as flowbuilder.  I'll fix
it next release I guess.