XDD Fix restart output
The restart update is determined fro the smallest offset in the
target offset table.  The target offset table was updated in
the post IO stats by the worker threads.  However, Tom modified
the update to the table to use it to sequentialize thread release
for loose and serial ordering.

Unfortunately, this means the TOT was not updated when NO ordering
was used.  I modified the post IO status to update the TOT if it
was not previously updated.  This is a fairly crappy solution,
but it gets restart working again.  The TOT needs massive revision.

The second issue with restart was that on a restart, XDD may erase
the contents of the restart file, or output offset 0, even though a
higher offset was supplied by the user.  I changed the behavior
so that a restart offset smaller than the user supplied is never
output.