XDDMCS status message improvement
This goes to a lot fo trouble to print accurate detailed status
messages for ongoing transfers.  It is similar, but not exactly
like SCP's status message.  It tries to make a smart decision
for what resolution to show data with so that the user can
determine if progress is occurring ina varying speed
environment.