Source reorganization for Python
Since Python is going to become a big part of the XDD
infrastructure, I've moved it into the regular source tree.  For
simplicity, I will change the license on all of the Python to
the free-est license I can, though I suspect it will still need
to be GPL in the long run.