Check magic cookies between remote endpoints (#22)
The idea here is to ensure endpoints are properly paired by exchanging
a "cookie." If the source side cookie and destination side cookie
don't match then the transfer will fail immediately. Our inspiration
is XWindows and MIT-MAGIC-COOKIE-1.

This feature won't be useful until the cookie can be set by a command
line option.