Establish e2e connections in worker threads
Connection establishment was moved to the worker threads from the
target thread. This change will allow the workers to have different
connections in the future (a requirement for multinic). Currently we
still only support a single connection, i.e. one remote host.

A mutex protects each connection. The first worker to hold the mutex
will establish the connection. The other workers will use the
connection that was already established.