Add calientctl.py
This is a basic software controller for the Calient optical switch.
It adds, removes, refreshes, and shows the coross connections for
the Calient optical switch.  The session establishment stuff isn't
perfect, so sometimes the Calient will reject the session
(this typically only happens when you establish several new
sessions in a row).  Wireshark diagnoses that something is slightly
different about that HTTP GET, but I don't know what it is.  A quick
patch is welcome (it should just be a quick mod to the _renewSession
method).

Between this and anuectl, users should pretty much be able to
establish their own long distance circuits.