Get nightly make test to run
This change gets the nightly make test to work.  The key is that the file
test_config can be generated in 2 ways, by the nightly gen script or by
copying from an existing template.  If the test_config exists and is newer,
we don't want to copy a template over it.  Otoh, if neither test_config or a
valid template exists, flag an error during make.